FIX sobre player schema

+++ src/schemas/player_schema.py
@@ -1,13 +1,14 @@
 from typing import Optional
+from sqlalchemy import Float
 
 from pydantic import BaseModel, Field, ConfigDict
 
 class PlayerStatsUpdate(BaseModel):
-    punteria: Optional[int] = None
-    velocidad: Optional[int] = None
-    dribbling: Optional[int] = None
-    defensa: Optional[int] = None
-    magia: Optional[int] = None
+    punteria: Optional[float] = None
+    velocidad: Optional[float] = None
+    dribbling: Optional[float] = None
+    defensa: Optional[float] = None
+    magia: Optional[float] = None
 
 
 
@@ -15,12 +16,12 @@ class PlayerResponse(BaseModel):
     id: int
     name: str
     cant_partidos: int
-    elo: int
-    punteria: int
-    velocidad: int
-    dribbling: int
-    defensa: int
-    magia: int
+    elo: float
+    punteria: float
+    velocidad: float
+    dribbling: float
+    defensa: float
+    magia: float
 
     class Config:
         model_config = ConfigDict(from_attributes=True)  # reemplaza orm_mode=True
